workaround for euca-* tooling no_proxy issue
the euca-* tools (at least euca-describe-services)
not support no_proxy properly. so in an environment
where an http_proxy is specified, they will attempt
to use the proxy even if the target is explicitly
excluded by no_proxy.

in the case of euca-describe-services this results
in attempted queries to the proxy server for "localhost"

This workaround prevents this issue from stopping
the fasttrack tooling by removing all proxy environment
variables before chef runs any of the euca-* tools